refactor: replace macOS aws-cli hostname gate with HAS_AWS
Collapses outer LAPTOP||STUDIO||...||RATNA + inner MACOS check into
a single HAS_AWS && MACOS condition. Removes RATNA which was a
deprecated host. Reduces nesting by moving mkdir -p and printf
statements out of the redundant inner MACOS conditional.

+++ setup_env.sh
@@ -1002,17 +1002,15 @@ EOM
     fi
   fi
 
-  if [[ -n ${LAPTOP} ]] || [[ -n ${STUDIO} ]] || [[ -n ${RECEPTION} ]] || [[ -n ${OFFICE} ]] || [[ -n ${HOMES} ]] || [[ -n ${RATNA} ]]; then
+  if [[ -n ${HAS_AWS} ]] && [[ -n ${MACOS} ]]; then
     mkdir -p ${HOME}/software_downloads/awscli
-    if [[ -n ${MACOS} ]]; then
-      printf "Installing aws-cli on MacOS\\n"
-      if [[ ! -f ${HOME}/software_downloads/awscli/AWSCLIV2.pkg ]]; then
-        wget -O ${HOME}/software_downloads/awscli/AWSCLIV2.pkg "https://awscli.amazonaws.com/AWSCLIV2.pkg"
-        sudo installer -pkg ${HOME}/software_downloads/awscli/AWSCLIV2.pkg -target /
-        rm -f ${HOME}/software_downloads/awscli/AWSCLIV2.pkg
-        if [[ -x $(command -v aws) ]]; then
-          printf "aws-cli is installed MacOS\\n"
-        fi
+    printf "Installing aws-cli on MacOS\\n"
+    if [[ ! -f ${HOME}/software_downloads/awscli/AWSCLIV2.pkg ]]; then
+      wget -O ${HOME}/software_downloads/awscli/AWSCLIV2.pkg "https://awscli.amazonaws.com/AWSCLIV2.pkg"
+      sudo installer -pkg ${HOME}/software_downloads/awscli/AWSCLIV2.pkg -target /
+      rm -f ${HOME}/software_downloads/awscli/AWSCLIV2.pkg
+      if [[ -x $(command -v aws) ]]; then
+        printf "aws-cli is installed MacOS\\n"
       fi
     fi
   fi
